snapshot: mqtt implementation experiments
This is a partial implementation of a homie device node that deals with
exposing a read-only view of the device alarms.

Most notably I haven't used the `$alarm` mechanism for this as the
device has more nuance than just "firing" or "clear". It can also have a
state wherein the alarm was attempted to be cleared but is still firing,
or a state where some condition is being triggered but the time didn't
lapse for it to actually fire. A common alarm here is both the heater
overheat sensor and the rotor guard system which keeps track of the
rotation of the rotor.

Anyway, the plan is to use the `/set` for clearing and `$target` to
indicate that a clear has been attempted but not yet actually cleared.

For the time being the prototype node implementation builds up a stream
of polls up to the device, but those return just the modbus register
values right now. I will want to keep the logic around `set`s and
`$targets` and suchwhatnot inside the node implementation.