Add a node to control digital/virtual inputs
These are useful if you want to programatically enable one of the modes
available. I, for example, use it to trigger the cooker hood mode as
mine is not colocated with the HRU and I don't have a wire going between
them.

There should also be registers to configure actual physical inputs
(denoted as `UI?`,) but modbus documentation does not reveal which
addresses these are and so configuration here would require either
asking folks at SystemAIR nicely (I don't personally have a need for
these myself) or reverse-engineering it from the commands that the cloud
sends when the HRU is being configured.

Additionally this de-TLAs one of the alarms and also moves one of the
cooling recovery properties from cooler to heat-exchanger node.